kid-coder-carnival-api/kidcodercarnivalapi/requirements.txt
django>=3.0.0
psycopg2>=2.8.0
djangorestframework>=3.0.0
djangorestframework-simplejwt>=4.0.0
drf-yasg>=1.20.0
python-dotenv>=0.17.0
django-cors-headers
